Increases max-width of .main-background-img.
The original reason for the size of the main background image is lost to the ages, and it is unknown whether or not it's still relevant to the design anyway.

This update increases the size of the imgage slightly, in order to match the width of the main content area, which at least is a trackable reason.

It also prevents rogue users from uploading a background image that is wider than recommended.