Adds blur effect to future bg implementation.
These styles will be applied to a future background implementation in which a very small image is provided, blown up to size, and then blurred to heck.

After that, the large image is loaded, swapped in under the blur, and then quickly sharpened for a full-blown background experience.

This might wait until Drupal, or it might be sooner, but it's in the source for now.